DIV-5166: removing endpoint for making case eligible for DA since this is only supposed to be called by the job scheduler.
