+++ src/integrationTest/resources/application.properties
@@ -51,7 +51,6 @@ case.orchestration.maintenance.submit-dn.context-path=/submit-dn
 case.orchestration.retrieve-case.context-path=/retrieve-case
 case.orchestration.payment-update.context-path=/payment-update
 case.orchestration.amend-petition.context-path=/amend-petition
-case.orchestration.make-case-eligible-for-da.context-path=/make-case-eligible-for-da
 case.orchestration.jobScheduler.make-case-eligible-for-da.context-path=/cases/da/make-eligible
 case.orchestration.co-respondent-generate-answers.context-path=/co-respondent-generate-answers
 case.orchestration.update-bulk-pronouncement-date.context-path=/bulk/pronounce/submit
