+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/controller/OrchestrationController.java
@@ -29,7 +29,6 @@ import uk.gov.hmcts.reform.divorce.orchestration.domain.model.idam.UserDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.payment.PaymentUpdate;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
 import uk.gov.hmcts.reform.divorce.orchestration.service.CaseOrchestrationService;
-import uk.gov.hmcts.reform.divorce.orchestration.service.CaseOrchestrationServiceException;
 
 import java.util.Map;
 
@@ -355,30 +354,6 @@ public class OrchestrationController {
         throws WorkflowException {
 
         return ResponseEntity.ok(orchestrationService.amendPetition(caseId, authorizationToken));
-    }
-
-    @PostMapping(path = "/make-case-eligible-for-da/{caseId}",
-        produces = MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "Makes a case eligible for Decree Absolute")
-    @ApiResponses(value = {
-        @ApiResponse(code = 200, message = "Case was updated in CCD",
-            response = CaseResponse.class),
-        @ApiResponse(code = 400, message = "Bad Request")})
-    public ResponseEntity<Map<String, Object>> makeCaseEligibleForDecreeAbsolute(
-        @RequestHeader(value = "Authorization") String authorizationToken,
-        @PathVariable String caseId) {
-
-        ResponseEntity<Map<String, Object>> responseEntity;
-
-        try {
-            responseEntity = ResponseEntity.ok(orchestrationService.makeCaseEligibleForDA(authorizationToken, caseId));
-            log.info("Case id {} made eligible for DA.", caseId);
-        } catch (CaseOrchestrationServiceException e) {
-            responseEntity = ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-            log.error("Error occurred making case id {} eligible for DA.", caseId);
-        }
-
-        return responseEntity;
     }
 
 }
